% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_readmes.R
\name{find_READMES}
\alias{find_READMES}
\title{Find all READMES}
\usage{
find_READMES()
}
\arguments{
\item{all_readmes}{a vector of all readme file locations}
}
\description{
This is the first step, to find all the READMEs
}
