% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_search_text.R, R/get_all_readme_text.R
\name{find_search_text}
\alias{find_search_text}
\title{This search text}
\usage{
find_search_text(readme_text, search_string)

find_search_text(readme_text, search_string)
}
\arguments{
\item{readme_text}{a list output from get_all_readme_text, probably could be a class}

\item{search_string}{what you are looking for}
}
\value{
the subset of list elements with that text string

the subset of list elements with that text string
}
\description{
Obviously a lot to update here, but not a bad start

Obviously a lot to update here, but not a bad start
}
