% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_in_readmes.R
\name{get_all_readme_text}
\alias{get_all_readme_text}
\title{Get all text in the readmes}
\usage{
get_all_readme_text(all_readmes)
}
\arguments{
\item{all_readmes}{the file names for all readmes}
}
\value{
a loosely organized list of all text in each file, that you can search
}
\description{
Get all text. Exported for now, maybe not forever.
}
