% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_config.R
\name{set_config}
\alias{set_config}
\title{Sets configuration file location}
\usage{
set_config(config_value = NULL)
}
\arguments{
\item{config_path}{The full path to the getANCHORconfig.R file}
}
\description{
You only need to run this once. This has some limited error checking but assumes
you are passing in a valid path
}
